Create GameTask
This class is an example of additional study I undertook to add an extra element to my game. It uses the methods in the class TimerTask by inheriting from it, and uses them to create a custom timer class (GameTask) with functions for my game. This is used to time the player, and create a more challenging game overall.