Create Enterable
This interface Enterable is used to help implement outputs based on where the player can go. If the player walks in to a wall, as described in the Wall class, they will receive a different output than if the player walks in to a Room, as shown by the use of the void enter method in Room.